HOCS-2561 - Remove debug logging and relocate ComplaintData
