+++ src/main/java/uk/gov/digital/ho/hocs/queue/data/UKVIComplaintData.java
@@ -1,4 +1,4 @@
-package uk.gov.digital.ho.hocs.client;
+package uk.gov.digital.ho.hocs.queue.data;
 
 import com.jayway.jsonpath.JsonPath;
 import com.jayway.jsonpath.PathNotFoundException;
