+++ src/test/java/uk/gov/digital/ho/hocs/queue/UKVIComplaintServiceTest.java
@@ -5,7 +5,7 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.digital.ho.hocs.application.ClientContext;
-import uk.gov.digital.ho.hocs.client.UKVIComplaintData;
+import uk.gov.digital.ho.hocs.queue.data.UKVIComplaintData;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
